# contracts/feedback/openapi.yaml
openapi: 3.1.0
...
info:
  title: Feedback Service API
  version: 1.0.0-alpha.1
  description: |
    ホスト業務システムから独立したフィードバック基盤の公開契約。
    Web GIS の project、GIS、業務 CRUD、タイル API を含めない。
servers:
  - url: /feedback/v1
security:
  - bearerAuth: []
...
paths:
...
  /exports:
    post:
      tags: [administration]
      operationId: createFeedbackExport
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackExportRequest"
      responses:
        "202":
          description: 受付済み export job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackExportJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /exports/{exportId}:
    parameters:
      - $ref: "#/components/parameters/ExportId"
    get:
      tags: [administration]
      operationId: getFeedbackExport
      responses:
        "200":
          description: export job の現在状態
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackExportJob"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /exports/{exportId}/download:
    parameters:
      - $ref: "#/components/parameters/ExportId"
    get:
      tags: [administration]
      operationId: downloadFeedbackExport
      responses:
        "200":
          description: 認可済みで保存期限内のCSV/XLSX
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            text/csv:
              schema: { type: string, format: binary }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /backup-policy:
    get:
      tags: [administration]
      operationId: getFeedbackBackupPolicy
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
      responses:
        "200":
          description: workspaceの自動証跡バックアップ方針
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackBackupPolicyView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      tags: [administration]
      operationId: patchFeedbackBackupPolicy
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/FeedbackBackupPolicy" }
      responses:
        "200":
          description: 更新した自動バックアップ方針
          headers:
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackBackupPolicyView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /backups:
    get:
      tags: [administration]
      operationId: listFeedbackBackups
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: 自動バックアップ実行履歴
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackBackupRunPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /backups/{backupId}:
    parameters:
      - $ref: "#/components/parameters/BackupId"
    get:
      tags: [administration]
      operationId: getFeedbackBackup
      responses:
        "200":
          description: 自動バックアップ実行状態
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackBackupRun" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /backups/{backupId}/download:
    parameters:
      - $ref: "#/components/parameters/BackupId"
    get:
      tags: [administration]
      operationId: downloadFeedbackBackup
      responses:
        "200":
          description: manifest・CSV・証跡画像を含む自己完結ZIP
          headers:
            Content-Disposition:
              schema: { type: string }
            ETag:
              schema: { type: string }
          content:
            application/zip:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /backups/{backupId}/retry:
    parameters:
      - $ref: "#/components/parameters/BackupId"
    post:
      tags: [administration]
      operationId: retryFeedbackBackup
      parameters:
        - $ref: "#/components/parameters/ApplicationKeyQuery"
        - $ref: "#/components/parameters/ExternalWorkspaceKeyQuery"
      responses:
        "200":
          description: 再試行待ちへ戻したバックアップ
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FeedbackBackupRun" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        直接OIDC JWTまたは契約済みtoken exchange JWT。直接OIDC JWTは検証対象のissuer、audience、
        時刻、設定済みsubject claimに加え、`feedback_permissions`を1件以上のFeedbackPermissionからなる
        文字列配列として必須とする。直接OIDCのpermissionは全resourceでDB membershipと交差する。
        token exchange JWTのresource scopeとclaimはtoken exchange公開契約に従う。
  parameters:
...
    ApplicationKeyQuery:
      name: applicationKey
      in: query
      required: true
      schema: { type: string, pattern: "^[a-z][a-z0-9-]{0,62}$" }
...
    ExternalWorkspaceKeyQuery:
      name: externalWorkspaceKey
      in: query
      required: true
      schema: { type: string, minLength: 1, maxLength: 200 }
...
    ExportId:
      name: exportId
      in: path
      required: true
      schema: { type: string, format: uuid }
    BackupId:
      name: backupId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 2000 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, minLength: 3, maxLength: 200 }
...
  headers:
    ETag:
      description: resource version。更新時は If-Match へそのまま指定する。
      schema: { type: string }
  responses:
    BadRequest:
      description: 契約または manifest に適合しない入力
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Unauthorized:
      description: 未認証または token 無効
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Forbidden:
      description: feedback permission 不足
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    NotFound:
      description: resource 不在または存在秘匿
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    Conflict:
      description: open session 競合などの一意制約違反
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    PreconditionFailed:
      description: ETag が古い
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
...
    TooManyRequests:
      description: rate limit または quota 超過
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
    ServiceUnavailable:
      description: DB または必須依存が利用不能
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/FeedbackProblem" }
...
  schemas:
...
    FeedbackProblem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code: { type: string, pattern: "^[a-z0-9_.-]+$" }
        requestId: { type: string, minLength: 1, maxLength: 200 }
...
    FeedbackExportRequest:
      type: object
      additionalProperties: false
      required: [applicationKey, environmentKey, externalWorkspaceKey, format]
      properties:
        applicationKey: { type: string }
        environmentKey: { type: string }
        externalWorkspaceKey: { type: string }
        sessionId: { type: [string, "null"], format: uuid }
        format: { enum: [csv, xlsx, evidence-package] }
        locale: { type: string, default: ja-JP }
        timezone: { type: string, default: Asia/Tokyo }
    FeedbackExportJob:
      type: object
      additionalProperties: false
      required: [id, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        status: { enum: [queued, running, completed, failed] }
        downloadUrl: { type: [string, "null"], format: uri-reference }
        expiresAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        error: { type: [string, "null"], maxLength: 2000 }
    FeedbackBackupPolicy:
      type: object
      additionalProperties: false
      required: [enabled, timezone, fullBackupAt, incrementalIntervalMinutes, includeEvidence, retentionDays]
      properties:
        enabled: { type: boolean, default: false }
        timezone: { type: string, default: Asia/Tokyo, maxLength: 100 }
        fullBackupAt: { type: string, pattern: "^[0-2][0-9]:[0-5][0-9]$", default: "02:00" }
        incrementalIntervalMinutes: { type: integer, minimum: 15, maximum: 1440, default: 60 }
        includeEvidence: { type: boolean, default: true }
        retentionDays: { type: [integer, "null"], minimum: 1, maximum: 3650 }
    FeedbackBackupPolicyView:
      type: object
      additionalProperties: false
      required: [policy, nextExecutionAt, nextFullAt, nextIncrementalAt, lastSuccessfulAt, changeCursor, auditCursor]
      properties:
        policy: { $ref: "#/components/schemas/FeedbackBackupPolicy" }
        nextExecutionAt: { type: [string, "null"], format: date-time }
        nextFullAt: { type: [string, "null"], format: date-time }
        nextIncrementalAt: { type: [string, "null"], format: date-time }
        lastSuccessfulAt: { type: [string, "null"], format: date-time }
        changeCursor: { type: integer, minimum: 0 }
        auditCursor: { type: integer, minimum: 0 }
    FeedbackBackupRun:
      type: object
      additionalProperties: false
      required: [id, kind, status, scheduledFor, fromChangeSequence, fromAuditSequence, historyCoverageStartedAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        kind: { enum: [full, incremental] }
        status: { enum: [queued, running, completed, failed, superseded] }
        scheduledFor: { type: string, format: date-time }
        downloadUrl: { type: [string, "null"], format: uri-reference }
        fromChangeSequence: { type: integer, minimum: 0 }
        toChangeSequence: { type: [integer, "null"], minimum: 0 }
        fromAuditSequence: { type: integer, minimum: 0 }
        toAuditSequence: { type: [integer, "null"], minimum: 0 }
        archiveSha256: { type: [string, "null"], pattern: "^[a-f0-9]{64}$" }
        archiveBytes: { type: [integer, "null"], minimum: 0 }
        entryCounts:
          type: [object, "null"]
          additionalProperties: { type: integer, minimum: 0 }
        historyCoverageStartedAt: { type: string, format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }
        completedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        error: { type: [string, "null"], maxLength: 2000 }
    FeedbackBackupRunPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FeedbackBackupRun" }
        nextCursor: { type: [string, "null"] }
...